I added a test for the href as well as clicking all of the buttons in register_test.py as well as test_login.py
