+++ tests/test_login.py
@@ -22,6 +22,16 @@ class Test(unittest.TestCase):
         self.driver.find_element(By.ID,"password").send_keys(password)
         self.driver.find_element(By.ID,"terms-privacy-check").click()
         self.driver.find_element(By.ID,"login").click()
+    def test_click(self):
+        WebDriverWait(self.driver, 10).until(EC.presence_of_element_located((By.ID, "email")))
+        btns = self.driver.find_elements("xpath","//div[contains(@class, 'col btn btn-secondary button me-1')]")
+        for btn in btns:
+            btn.click()
+    def test_href(self):
+        WebDriverWait(self.driver, 10).until(EC.presence_of_element_located((By.ID, "email")))
+        links = self.driver.find_elements("xpath","//div[contains(@href, '#')]")
+        for link in links:
+            link.click()
     def tearDown(self):
         self.driver.close()
 if __name__=="__main__":
